Python/requirements.txt
anyio==4.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
confluent-kafka==2.3.0
dnspython==2.4.2
grpcio==1.60.0
grpcio-tools==1.60.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
mongoengine==0.27.0
protobuf==4.25.1
pymongo==4.6.1
python-telegram-bot==20.7
requests==2.31.0
setuptools==69.0.2
sniffio==1.3.0
style==1.1.0
urllib3==2.1.0
